+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/trash/WikiPageTrashHandler.java
@@ -27,17 +27,16 @@ import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.trash.TrashActionKeys;
 import com.liferay.portal.kernel.trash.TrashHandler;
-import com.liferay.portal.kernel.trash.TrashHandlerRegistryUtil;
 import com.liferay.portal.kernel.trash.TrashRenderer;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
+import com.liferay.trash.TrashHelper;
 import com.liferay.trash.kernel.exception.RestoreEntryException;
 import com.liferay.trash.kernel.model.TrashEntry;
 import com.liferay.trash.kernel.model.TrashEntryConstants;
-import com.liferay.trash.kernel.util.TrashUtil;
 import com.liferay.wiki.asset.WikiPageAssetRenderer;
 import com.liferay.wiki.constants.WikiPortletKeys;
 import com.liferay.wiki.engine.impl.WikiEngineRenderer;
@@ -240,44 +239,17 @@ public class WikiPageTrashHandler extends BaseWikiTrashHandler {
 	}
 
 	@Override
-	public List<TrashRenderer> getTrashContainerModelTrashRenderers(
-			long classPK, int start, int end)
-		throws PortalException {
-
-		List<TrashRenderer> trashRenderers = new ArrayList<>();
-
-		WikiPage page = _wikiPageLocalService.getPage(classPK);
-
-		List<WikiPage> pages = _wikiPageLocalService.getChildren(
-			page.getNodeId(), true, page.getTitle(),
-			WorkflowConstants.STATUS_IN_TRASH);
-
-		for (WikiPage curPage : pages) {
-			TrashHandler trashHandler =
-				TrashHandlerRegistryUtil.getTrashHandler(
-					WikiPage.class.getName());
-
-			TrashRenderer trashRenderer = trashHandler.getTrashRenderer(
-				curPage.getResourcePrimKey());
-
-			trashRenderers.add(trashRenderer);
-		}
-
-		return trashRenderers;
-	}
-
-	@Override
 	public TrashedModel getTrashedModel(long classPK) {
 		return _wikiPageLocalService.fetchLatestPage(
 			classPK, WorkflowConstants.STATUS_ANY, false);
 	}
 
 	@Override
-	public List<TrashRenderer> getTrashModelTrashRenderers(
+	public List<TrashedModel> getTrashModelTrashedModels(
 			long classPK, int start, int end, OrderByComparator obc)
 		throws PortalException {
 
-		List<TrashRenderer> trashRenderers = new ArrayList<>();
+		List<TrashedModel> trashedModels = new ArrayList<>();
 
 		WikiPage page = _wikiPageLocalService.getPage(classPK);
 
@@ -286,17 +258,10 @@ public class WikiPageTrashHandler extends BaseWikiTrashHandler {
 			WorkflowConstants.STATUS_IN_TRASH, start, end, obc);
 
 		for (WikiPage curPage : pages) {
-			TrashHandler trashHandler =
-				TrashHandlerRegistryUtil.getTrashHandler(
-					WikiPage.class.getName());
-
-			TrashRenderer trashRenderer = trashHandler.getTrashRenderer(
-				curPage.getResourcePrimKey());
-
-			trashRenderers.add(trashRenderer);
+			trashedModels.add(curPage);
 		}
 
-		return trashRenderers;
+		return trashedModels;
 	}
 
 	@Override
@@ -304,7 +269,8 @@ public class WikiPageTrashHandler extends BaseWikiTrashHandler {
 		WikiPage page = _wikiPageLocalService.getLatestPage(
 			classPK, WorkflowConstants.STATUS_ANY, false);
 
-		return new WikiPageAssetRenderer(page, _wikiEngineRenderer);
+		return new WikiPageAssetRenderer(
+			page, _wikiEngineRenderer, _trashHelper);
 	}
 
 	@Override
@@ -432,7 +398,7 @@ public class WikiPageTrashHandler extends BaseWikiTrashHandler {
 				checkRestorableEntry(
 					curPage.getResourcePrimKey(), 0, containerModelId,
 					curPage.getTitle(),
-					TrashUtil.getOriginalTitle(curPage.getTitle()));
+					_trashHelper.getOriginalTitle(curPage.getTitle()));
 			}
 		}
 	}
@@ -507,6 +473,9 @@ public class WikiPageTrashHandler extends BaseWikiTrashHandler {
 
 	@Reference
 	private Portal _portal;
+
+	@Reference
+	private TrashHelper _trashHelper;
 
 	private WikiEngineRenderer _wikiEngineRenderer;
 	private WikiPageLocalService _wikiPageLocalService;