+++ modules/apps/foundation/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/result/display/builder/SearchResultSummaryDisplayBuilder.java
@@ -25,6 +25,7 @@ import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Indexer;
+import com.liferay.portal.kernel.search.IndexerRegistry;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.search.Summary;
@@ -72,11 +73,10 @@ public class SearchResultSummaryDisplayBuilder {
 
 	public SearchResultSummaryDisplayContext build() throws Exception {
 		String className = _document.get(Field.ENTRY_CLASS_NAME);
-		long classPK = GetterUtil.getLong(_document.get(Field.ENTRY_CLASS_PK));
+		long classPK = getEntryClassPK();
 
 		AssetRendererFactory<?> assetRendererFactory =
-			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
-				className);
+			getAssetRendererFactoryByClassName(className);
 
 		AssetRenderer<?> assetRenderer = null;
 
@@ -110,6 +110,12 @@ public class SearchResultSummaryDisplayBuilder {
 		_assetEntryLocalService = assetEntryLocalService;
 	}
 
+	public void setAssetRendererFactoryLookup(
+		AssetRendererFactoryLookup assetRendererFactoryLookup) {
+
+		_assetRendererFactoryLookup = assetRendererFactoryLookup;
+	}
+
 	public void setCurrentURL(String currentURL) {
 		_currentURL = currentURL;
 	}
@@ -124,6 +130,10 @@ public class SearchResultSummaryDisplayBuilder {
 
 	public void setImageRequested(boolean imageRequested) {
 		_imageRequested = imageRequested;
+	}
+
+	public void setIndexerRegistry(IndexerRegistry indexerRegistry) {
+		_indexerRegistry = indexerRegistry;
 	}
 
 	public void setLanguage(Language language) {
@@ -172,6 +182,12 @@ public class SearchResultSummaryDisplayBuilder {
 		_searchResultPreferences = searchResultPreferences;
 	}
 
+	public void setSearchResultViewURLSupplier(
+		SearchResultViewURLSupplier searchResultViewURLSupplier) {
+
+		_searchResultViewURLSupplier = searchResultViewURLSupplier;
+	}
+
 	public void setThemeDisplay(ThemeDisplay themeDisplay) {
 		_themeDisplay = themeDisplay;
 	}
@@ -211,7 +227,8 @@ public class SearchResultSummaryDisplayBuilder {
 		buildImage(searchResultSummaryDisplayContext, className, classPK);
 		buildLocaleReminder(searchResultSummaryDisplayContext, summary);
 		buildModelResource(searchResultSummaryDisplayContext, className);
-		buildUserPortrait(searchResultSummaryDisplayContext, assetEntry);
+		buildUserPortrait(
+			searchResultSummaryDisplayContext, assetEntry, className);
 		buildViewURL(className, classPK, searchResultSummaryDisplayContext);
 
 		return searchResultSummaryDisplayContext;
@@ -411,7 +428,16 @@ public class SearchResultSummaryDisplayBuilder {
 
 	protected void buildUserPortrait(
 		SearchResultSummaryDisplayContext searchResultSummaryDisplayContext,
-		AssetEntry assetEntry) {
+		AssetEntry assetEntry, String className) {
+
+		long entryClassPK = getEntryClassPK();
+
+		AssetEntry childAssetEntry = _assetEntryLocalService.fetchEntry(
+			className, entryClassPK);
+
+		if (childAssetEntry != null) {
+			assetEntry = childAssetEntry;
+		}
 
 		if (assetEntry != null) {
 			searchResultSummaryDisplayContext.setAssetEntryUserId(
@@ -424,9 +450,7 @@ public class SearchResultSummaryDisplayBuilder {
 		String className, long classPK,
 		SearchResultSummaryDisplayContext searchResultSummaryDisplayContext) {
 
-		String viewURL = SearchUtil.getSearchResultViewURL(
-			_renderRequest, _renderResponse, className, classPK,
-			_searchResultPreferences.isViewInContext(), _currentURL);
+		String viewURL = getSearchResultViewURL(className, classPK);
 
 		searchResultSummaryDisplayContext.setViewURL(viewURL);
 	}
@@ -454,13 +478,47 @@ public class SearchResultSummaryDisplayBuilder {
 		return assetEntry.getUserId();
 	}
 
+	protected AssetRendererFactory<?> getAssetRendererFactoryByClassName(
+		String className) {
+
+		if (_assetRendererFactoryLookup != null) {
+			return _assetRendererFactoryLookup.
+				getAssetRendererFactoryByClassName(className);
+		}
+
+		return AssetRendererFactoryRegistryUtil.
+			getAssetRendererFactoryByClassName(className);
+	}
+
+	protected long getEntryClassPK() {
+		return GetterUtil.getLong(_document.get(Field.ENTRY_CLASS_PK));
+	}
+
+	protected Indexer<Object> getIndexer(String className) {
+		if (_indexerRegistry != null) {
+			return _indexerRegistry.getIndexer(className);
+		}
+
+		return IndexerRegistryUtil.getIndexer(className);
+	}
+
+	protected String getSearchResultViewURL(String className, long classPK) {
+		if (_searchResultViewURLSupplier != null) {
+			return _searchResultViewURLSupplier.getSearchResultViewURL();
+		}
+
+		return SearchUtil.getSearchResultViewURL(
+			_renderRequest, _renderResponse, className, classPK,
+			_searchResultPreferences.isViewInContext(), _currentURL);
+	}
+
 	protected Summary getSummary(
 			String className, AssetRenderer<?> assetRenderer)
 		throws SearchException {
 
 		Summary summary = null;
 
-		Indexer<?> indexer = IndexerRegistryUtil.getIndexer(className);
+		Indexer indexer = getIndexer(className);
 
 		if (indexer != null) {
 			String snippet = _document.get(Field.SNIPPET);
@@ -553,10 +611,12 @@ public class SearchResultSummaryDisplayBuilder {
 
 	private boolean _abridged;
 	private AssetEntryLocalService _assetEntryLocalService;
+	private AssetRendererFactoryLookup _assetRendererFactoryLookup;
 	private String _currentURL;
 	private Document _document;
 	private boolean _highlightEnabled;
 	private boolean _imageRequested;
+	private IndexerRegistry _indexerRegistry;
 	private Language _language;
 	private Locale _locale;
 	private PortletURLFactory _portletURLFactory;
@@ -568,6 +628,7 @@ public class SearchResultSummaryDisplayBuilder {
 	private Stream<SearchResultImageContributor>
 		_searchResultImageContributorsStream = Stream.empty();
 	private SearchResultPreferences _searchResultPreferences;
+	private SearchResultViewURLSupplier _searchResultViewURLSupplier;
 	private ThemeDisplay _themeDisplay;
 
 }