+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaCombineLinesCheck.java
@@ -23,7 +23,6 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.ToolsUtil;
 import com.liferay.source.formatter.checks.util.JavaSourceUtil;
 
-import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -31,11 +30,6 @@ import java.util.regex.Pattern;
  * @author Hugo Huijser
  */
 public class JavaCombineLinesCheck extends BaseFileCheck {
-
-	public JavaCombineLinesCheck(List<String> excludes, int maxLineLength) {
-		_excludes = excludes;
-		_maxLineLength = maxLineLength;
-	}
 
 	@Override
 	protected String doProcess(
@@ -63,7 +57,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 
 				int lineLength = getLineLength(line);
 
-				if (lineLength > _maxLineLength) {
+				if (lineLength > getMaxLineLength()) {
 					previousLine = line;
 
 					continue;
@@ -234,7 +228,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 				replacement, new String[] {",\n", "\n"},
 				new String[] {StringPool.COMMA_AND_SPACE, StringPool.BLANK});
 
-			if (getLineLength(replacement) <= _maxLineLength) {
+			if (getLineLength(replacement) <= getMaxLineLength()) {
 				return _getCombinedLinesContent(
 					StringUtil.replace(content, match, replacement), pattern);
 			}
@@ -299,7 +293,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 				line, linePart, StringPool.BLANK);
 		}
 		else {
-			if (((linePart.length() + lineLength) <= _maxLineLength) &&
+			if (((linePart.length() + lineLength) <= getMaxLineLength()) &&
 				(line.endsWith(StringPool.OPEN_CURLY_BRACE) ||
 				 line.endsWith(StringPool.SEMICOLON))) {
 
@@ -334,7 +328,8 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 		int lineTabCount, int previousLineTabCount) {
 
 		if (Validator.isNull(line) || Validator.isNull(previousLine) ||
-			isExcludedPath(_excludes, absolutePath, lineCount)) {
+			isExcludedPath(
+				_FIT_ON_SINGLE_LINE_EXCLUDES, absolutePath, lineCount)) {
 
 			return null;
 		}
@@ -395,7 +390,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 
 		int previousLineLength = getLineLength(previousLine);
 
-		if ((trimmedLine.length() + previousLineLength) < _maxLineLength) {
+		if ((trimmedLine.length() + previousLineLength) < getMaxLineLength()) {
 			if (trimmedPreviousLine.startsWith("for ") &&
 				previousLine.endsWith(StringPool.COLON) &&
 				line.endsWith(StringPool.OPEN_CURLY_BRACE)) {
@@ -517,7 +512,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 			}
 		}
 
-		if ((trimmedLine.length() + previousLineLength) <= _maxLineLength) {
+		if ((trimmedLine.length() + previousLineLength) <= getMaxLineLength()) {
 			if (previousLine.endsWith(StringPool.OPEN_PARENTHESIS) &&
 				line.endsWith(") {") && (getLevel(line) < 0)) {
 
@@ -596,7 +591,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 					if ((previousChar != CharPool.CLOSE_PARENTHESIS) &&
 						(previousChar != CharPool.OPEN_PARENTHESIS) &&
 						(previousChar != CharPool.SPACE) &&
-						(previousLineLength + 1 + x) < _maxLineLength) {
+						(previousLineLength + 1 + x) < getMaxLineLength()) {
 
 						String linePart = trimmedLine.substring(0, x + 1);
 
@@ -627,7 +622,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 				x = trimmedLine.indexOf(" +", x + 1);
 
 				if ((x == -1) ||
-					(previousLineLength + 3 + x) > _maxLineLength) {
+					(previousLineLength + 3 + x) > getMaxLineLength()) {
 
 					break;
 				}
@@ -687,7 +682,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 				}
 
 				if (x != -1) {
-					while ((previousLineLength + 1 + x) < _maxLineLength) {
+					while ((previousLineLength + 1 + x) < getMaxLineLength()) {
 						String linePart = trimmedLine.substring(0, x + 1);
 
 						if (!ToolsUtil.isInsideQuotes(trimmedLine, x) &&
@@ -719,7 +714,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 					}
 				}
 				else if ((trimmedLine.length() + previousLineLength) <
-							_maxLineLength) {
+							getMaxLineLength()) {
 
 					if (!trimmedLine.startsWith("new ") ||
 						!line.endsWith(StringPool.OPEN_CURLY_BRACE)) {
@@ -745,7 +740,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 			int x = trimmedLine.indexOf(" = ");
 
 			if ((x != -1) && !ToolsUtil.isInsideQuotes(trimmedLine, x) &&
-				((previousLineLength + 2 + x) < _maxLineLength)) {
+				((previousLineLength + 2 + x) < getMaxLineLength())) {
 
 				String linePart = trimmedLine.substring(0, x + 3);
 
@@ -755,7 +750,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 			}
 			else if (trimmedLine.endsWith(" =") &&
 					 ((trimmedLine.length() + previousLineLength) <
-						 _maxLineLength)) {
+						 getMaxLineLength())) {
 
 				for (int i = 0;; i++) {
 					String nextLine = getLine(content, lineCount + i + 1);
@@ -795,7 +790,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 			if (linePart2.matches("[!=<>\\+\\-\\*]+ .*")) {
 				int y = trimmedLine.indexOf(StringPool.SPACE, x + 2);
 
-				if ((previousLineLength + y) <= _maxLineLength) {
+				if ((previousLineLength + y) <= getMaxLineLength()) {
 					return _getCombinedLinesContent(
 						content, line, trimmedLine, lineLength, lineCount,
 						previousLine, trimmedLine.substring(0, y), true, true,
@@ -827,7 +822,7 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 			}
 		}
 
-		if ((trimmedLine.length() + previousLineLength) > _maxLineLength) {
+		if ((trimmedLine.length() + previousLineLength) > getMaxLineLength()) {
 			return null;
 		}
 
@@ -858,6 +853,9 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 		return null;
 	}
 
+	private static final String _FIT_ON_SINGLE_LINE_EXCLUDES =
+		"fit.on.single.line.excludes";
+
 	private final Pattern _combinedLinesPattern1 = Pattern.compile(
 		"\n(\t*).+(=|\\]) (\\{)\n");
 	private final Pattern _combinedLinesPattern2 = Pattern.compile(
@@ -866,7 +864,5 @@ public class JavaCombineLinesCheck extends BaseFileCheck {
 		"(\n\t*(private|protected|public) void)\n\t+(\\w+\\(\\)( \\{)?\n)");
 	private final Pattern _combinedLinesPattern4 = Pattern.compile(
 		"(\n\t*(extends|implements))\n\t+([\\w.]+ \\{\n)");
-	private final List<String> _excludes;
-	private final int _maxLineLength;
 
 }